chore[litemall-admin]: 把用户相关的api集中到单一的user.js
