--- litemall-admin/src/api/address.js
@@ -1,9 +1,0 @@
-import request from '@/utils/request'
-
-export function listAddress(query) {
-  return request({
-    url: '/address/list',
-    method: 'get',
-    params: query
-  })
-}
