--- litemall-admin/src/api/collect.js
@@ -1,9 +1,0 @@
-import request from '@/utils/request'
-
-export function listCollect(query) {
-  return request({
-    url: '/collect/list',
-    method: 'get',
-    params: query
-  })
-}
