--- litemall-admin/src/api/feedback.js
@@ -1,9 +1,0 @@
-import request from '@/utils/request'
-
-export function listFeedback(query) {
-  return request({
-    url: '/feedback/list',
-    method: 'get',
-    params: query
-  })
-}
