--- litemall-admin/src/api/footprint.js
@@ -1,9 +1,0 @@
-import request from '@/utils/request'
-
-export function listFootprint(query) {
-  return request({
-    url: '/footprint/list',
-    method: 'get',
-    params: query
-  })
-}
