--- litemall-admin/src/api/history.js
@@ -1,9 +1,0 @@
-import request from '@/utils/request'
-
-export function listHistory(query) {
-  return request({
-    url: '/history/list',
-    method: 'get',
-    params: query
-  })
-}
