+++ litemall-admin/src/api/user.js
@@ -31,3 +31,44 @@ export function updateUser(data) {
     data
   })
 }
+
+export function listAddress(query) {
+  return request({
+    url: '/address/list',
+    method: 'get',
+    params: query
+  })
+}
+
+export function listCollect(query) {
+  return request({
+    url: '/collect/list',
+    method: 'get',
+    params: query
+  })
+}
+
+export function listFeedback(query) {
+  return request({
+    url: '/feedback/list',
+    method: 'get',
+    params: query
+  })
+}
+
+export function listFootprint(query) {
+  return request({
+    url: '/footprint/list',
+    method: 'get',
+    params: query
+  })
+}
+
+export function listHistory(query) {
+  return request({
+    url: '/history/list',
+    method: 'get',
+    params: query
+  })
+}
+
